#15 : Added ProductAssignment functionality with frontend and backend changes

See merge request internal_team/inventorymanagement!20

+++ src/main/java/com/poc/InventoryManagement/entity/Product.java
@@ -28,7 +28,7 @@ public class Product {
     @Column(name = "location")
     private String location;
     @Column(name = "owner")
-    private int owner;
+    private Long owner;
     @Column(name = "comments")
     private String comments;
     @Column(name = "trackingid")
